backend/app/routers/weather.py
from fastapi import APIRouter, HTTPException, Query
from pydantic import BaseModel
from typing import Dict, List, Optional, Any
import logging
import requests
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

router = APIRouter()
logger = logging.getLogger(__name__)

# You would typically store this in .env file
WEATHER_API_KEY = os.getenv("WEATHER_API_KEY", "")

class WeatherResponse(BaseModel):
    location: str
    current_conditions: Dict[str, Any]
    forecast: List[Dict[str, Any]]

@router.get("/weather", response_model=WeatherResponse)
async def get_weather(
    latitude: float = Query(..., description="Latitude coordinate"),
    longitude: float = Query(..., description="Longitude coordinate")
):
    """
    Get current weather and forecast for a specific location
    """
    try:
        # This is a placeholder for integrating with a real weather API
        # You would typically use something like OpenWeatherMap, WeatherAPI, etc.
        
        if not WEATHER_API_KEY:
            logger.warning("Weather API key not found in environment variables")
            # Return mock data for demonstration
            return {
                "location": f"Location at {latitude}, {longitude}",
                "current_conditions": {
                    "temperature": 25.5,
                    "humidity": 45,
                    "wind_speed": 15,
                    "precipitation": 0,
                    "weather": "Clear"
                },
                "forecast": [
                    {
                        "date": "2025-05-19",
                        "temperature_high": 27.8,
                        "temperature_low": 15.2,
                        "humidity": 40,
                        "precipitation_chance": 10,
                        "wind_speed": 12,
                        "weather": "Sunny"
                    },
                    {
                        "date": "2025-05-20",
                        "temperature_high": 29.1,
                        "temperature_low": 16.3,
                        "humidity": 38,
                        "precipitation_chance": 5,
                        "wind_speed": 14,
                        "weather": "Sunny"
                    }
                ]
            }
        
        # In a real implementation, you would call a weather API here
        # Example with OpenWeatherMap:
        # url = f"https://api.openweathermap.org/data/2.5/onecall?lat={latitude}&lon={longitude}&appid={WEATHER_API_KEY}&units=metric"
        # response = requests.get(url)
        # data = response.json()
        
        # Process and return the data
        # ...
        
    except Exception as e:
        logger.error(f"Error getting weather data: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))
