backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
joblib==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
tensorflow==2.15.0
kagglehub==0.2.8
matplotlib==3.8.0
pillow==10.0.1
seaborn==0.13.0
python-multipart==0.0.6
